updates to event interface

+++ Source/Owl.h
@@ -22,7 +22,7 @@
 #endif /* abs */
 
    void audioCallback(int16_t *src, int16_t *dst);
-   void setButton(uint8_t bid, uint8_t state);
+   void setButton(uint8_t bid, uint16_t state);
    void setParameter(uint8_t pid, uint16_t value);
    uint16_t getParameterValue(uint8_t index);
    void setup(); // main OWL setup
